Create inital profiles page
Currently a new profile is added to the db by start the application. In the future we must do that by fill some input fields and add a new profile by saving these information.
In addition we do not need checkboxes. We need a posibility to add a new profile and to select an existing profile. This is the first step by using the finance application.