Fix issue #13
I have create a class "ActivityMemory" where we can save the current profile.
We need this instance once because we need it for the transactions, too.
The error is fixed and the ProfileBean is saved in a common class.

Maybe we can change the name of the class. I have now idea how to name the class.